remove plotting_tools
